Logging: Waits to open logfile until first log is made.
TESTING=first unittest.

+++ log-test.py
@@ -10,15 +10,13 @@ import Log
 
 class LogTestCase(unittest.TestCase):
     HEADER_BYTES = 125
-    ENTRY_BYTES = 71
+    ENTRY_BYTES = 91
     ENTRIES_PER_FILE = 3
     MAX_BYTES_FILE = (HEADER_BYTES +                      # For header.
                       (ENTRIES_PER_FILE * ENTRY_BYTES) +  # For entries.
                       (ENTRIES_PER_FILE - 1)) # For commas between entries.
     MAX_BYTES_TOTAL = MAX_BYTES_FILE * 6
     LOG_PATH = "TestLog"
-    # 412 Header + 4 
-    # 125 Header only
 
     def __init__(self, *args, **kwargs):
         self.counter = 10  # Hack: all entries will be same size since counter
@@ -31,19 +29,12 @@ class LogTestCase(unittest.TestCase):
     def tearDown(self):
         pass
 
-    # TODO: size file=header+3 entries; total_max = 6 files = 18 entries
-    # TODO: rmdir, += 8 entries =  8 entries: 3 3 2
-    # TODO:        += 4 entries = 12 entries: 3 3 H3 3
-    # TODO:        += 5 entries = 17 entries: 3 3 H3 3 3 2
-    # TODO:        += 6 entries = 23 entries: x 3 H3 3 3 H3 3 2
-
-    def test_log_empty_dir(self): # ALL tests must start with 'test'
+    def __engine(self, clean_start, entry_count, expected_file_count):
         executive = Executive.Executive()
         counter_start = self.counter
-        entry_count = 8 # XXX: param
         beyond_max_entry = counter_start + entry_count
 
-        if os.path.exists(self.LOG_PATH): # XXX: param
+        if clean_start and os.path.exists(self.LOG_PATH):
             shutil.rmtree(self.LOG_PATH)  # Start by deleting the previous log
         with Log.Log(executive = executive,
                      path = self.LOG_PATH,
@@ -56,11 +47,10 @@ class LogTestCase(unittest.TestCase):
 
         # Verify by reading the data back from the files.
 
-        assert len(os.listdir(self.LOG_PATH)) == 3
+        assert len(os.listdir(self.LOG_PATH)) == expected_file_count
         entries = []
         for filename in sorted(os.listdir(self.LOG_PATH)):
             filepath = os.path.join(self.LOG_PATH, filename)
-            print "\n-- %s --" % filepath
             with file(filepath, 'r') as f:
                 whole_file = f.read()
                 data = json.loads(whole_file)
@@ -74,6 +64,32 @@ class LogTestCase(unittest.TestCase):
         for entry in entries:
             assert entry == counter_start
             counter_start += 1
+
+    def test_log_empty_dir(self):
+        self.__engine(clean_start = True,
+                      entry_count = 8,
+                      expected_file_count = 3)
+
+    # TODO: need to account for size of header in entry count
+    # TODO: need to include pre-existing entries in verification run
+
+    # TODO:        += 4 entries = 12 entries: 3 3 H3 3
+    #def test_log_add_to_file(self):
+    #    self.__engine(clean_start = True,
+    #                  entry_count = 4,
+    #                  expected_file_count = 4)
+
+    # TODO:        += 5 entries = 17 entries: 3 3 H3 3 3 2
+    #def test_log_new_file_on_first_log(self):
+    #    self.__engine(clean_start = True,
+    #                  entry_count = 5,
+    #                  expected_file_count = 6)
+
+    # TODO:        += 6 entries = 23 entries: x 3 H3 3 3 H3 3 2
+    #def test_log_wrap_and_delete_file(self):
+    #    self.__engine(clean_start = True,
+    #                  entry_count = 6,
+    #                  expected_file_count = 7)
 
 if __name__ == '__main__':
     unittest.main() # runs all tests
